kubernetesRun Take 1
After pushing to the docker hub, the pod boots with a deployment.
The logo is broken, might need to setup a root_url or some intelligence
on React side so that when the site is loaded on a browser it checks
the active URL so that it works, but then again this is the first React
app so just need to understand stuff.